# hms-api-documents.yaml
openapi: 3.0.3
info:
  title: Hospital Management System - OpenAPI 3.0
  description: |-
    Tài liệu API của Hệ thống quản lý bệnh viện (Hospital Management System)
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /staff/supportStaff/{supportStaff_id}:
    get:
      tags:
        - supportStaff
      summary: Xem thông tin nhân viên hỗ trợ
      description: Xem thông tin nhân viên hỗ trợ theo id
      operationId: getSupportStaffById
      parameters:
        - name: supportStaff_id
          in: path
          description: "Id của nhân viên hỗ trợ cần xem chi tiết. Dùng T2g3cyzImX69qpWh5nuN để test"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OneSupportStaff"
            application/xml:
              schema:
                $ref: "#/components/schemas/OneSupportStaff"
        "400":
          description: Invalid supportStaff_id supplied
        "404":
          description: Order not found
    put:
      tags:
        - supportStaff
      summary: Cập nhật thông tin của nhân viên hỗ trợ
      description: Cập nhật thông tin của nhân viên hỗ trợ
      operationId: updateSupportStaffById
      parameters:
        - name: supportStaff_id
          in: path
          description: "Id của nhân viên hỗ trợ cần xem chi tiết. Dùng T2g3cyzImX69qpWh5nuN để test"
          required: true
          schema:
            type: string
      requestBody:
        description: Cập nhật thông tin của nhân viên hỗ trợ
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSupportStaff"
          application/xml:
            schema:
              $ref: "#/components/schemas/AddSupportStaff"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/AddSupportStaff"
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid supportstaff_id supplied
        "404":
          description: Doctor not found
    delete:
      tags:
        - supportStaff
      summary: Xóa nhân viên hỗ trợ
      description: Xóa nhân viên hỗ trợ khỏi cơ sở dữ liệu
      operationId: deleteSupportStaffById
      parameters:
        - name: supportStaff_id
          in: path
          description: Id của nhân viên hỗ trợ cần xoá. **Tạo 1 nhân viên hỗ trợ mới rồi lên firebase lấy id để xoá.**
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid supportStaff_id supplied
...
components:
  schemas:
...
    OneSupportStaff:
      type: object
      properties:
        id:
          type: string
          example: T2g3cyzImX69qpWh5nuN
        firstName:
          type: string
          example: Trump
        lastName:
          type: string
          example: Donald
        dateOfBirth:
          type: string
          format: date
          example: 12/06/2002
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "5601928374"
        absence:
          type: boolean
          example: false
    AddSupportStaff:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - salary
        - absence
      type: object
      properties:
        firstName:
          type: string
          example: SS
        lastName:
          type: string
          example: Nguyen
        dateOfBirth:
          type: string
          format: date
          example: 12/31/1685
        gender:
          type: string
          example: male
        phoneNumber:
          type: string
          example: "1224056829"
        salary:
          type: integer
          example: 294620556
        absence:
          type: boolean
          example: true
...